Recover mover phase timing

+++ source/Thunderball/Mover.h
@@ -44,7 +44,7 @@ public:
 
 	void Translate(float param_1, float param_2);
 	void SetPos(float param_1, float param_2);
-	float GetTimeTillPhase(int param_1, float param_2);
+	int GetTimeTillPhase(int param_1, float param_2);
 	float GetMovePos(int param_1);
 	void CalcPos(int param_1, float param_2);
 	void DrawMovement(Graphics* g);
